--- src/react-challenges/Command.tsx
@@ -1,3 +1,0 @@
-export function Command(){
-  return <input />
-}
