+++ src/react-challenges/types/types.ts
@@ -1,0 +1,8 @@
+import { Vim } from "../../vim/vim";
+
+export type Challenge = {
+  description: string;
+  content: string;
+  expected: string | ((vim: Vim) => boolean);
+  strokes: number;
+}
